added highlight color picker
changed filenames
made ColorPickerModal a separate file
