+++ src/components/ui/basics/TipTapEditor/TipTapEditor.tsx
@@ -65,7 +65,8 @@ import { Edit, FileExport } from 'tabler-icons-react';
 import { LinkButton, UnlinkButton } from './Components/LinkButton';
 import { ToggleCodeBlock } from './Components/ToggleCodeBlock';
 import { IProgrammingLanguage } from '@custom-types/data/tiptap';
-import { ColorPickerButton } from './Components/ColorPicker';
+import { ColorPickerButton } from './Components/ColorPickerButton';
+import { HighLightColorButton } from './Components/HighlightColorButton';
 
 const content =
   '<h2 style="text-align: center;">Welcome to Mantine rich text editor</h2><p>RichTextEditor component focuses on usability and is designed to be as simple as possible to bring a familiar editing experience to regular users. RichTextEditor is based on <a href="https://tiptap.dev/" rel="noopener noreferrer" target="_blank">Tiptap.dev</a> and supports all of its features:</p><ul><li>General text formatting: <strong>bold</strong>, <em>italic</em>, <u>underline</u>, <s>strike-through</s> </li><li>Headings (h1-h6)</li><li>Sub and super scripts (<sup>&lt;sup /&gt;</sup> and <sub>&lt;sub /&gt;</sub> tags)</li><li>Ordered and bullet lists</li><li>Text align&nbsp;</li><li>And all <a href="https://tiptap.dev/extensions" target="_blank" rel="noopener noreferrer">other extensions</a></li></ul>';
@@ -136,7 +137,7 @@ export const TipTapEditor = () => {
       Dropcursor,
       FloatingMenu,
       Heading,
-      Highlight,
+      Highlight.configure({ multicolor: true }),
       History,
       Image,
       Italic,
@@ -170,6 +171,7 @@ export const TipTapEditor = () => {
               <ToggleStrikethrough editor={editor} />
               <RichTextEditor.ClearFormatting />
               <ColorPickerButton editor={editor} />
+              <HighLightColorButton editor={editor} />
             </RichTextEditor.ControlsGroup>
             <RichTextEditor.ControlsGroup className={styles.toolbar_group}>
               <ToggleCodeBlock
